Thentos.Api

Overview

Thentos distinguishes between transactions and actions.

transactions are acidic in the sense used in acid-state, and live in ThentosQuery or ThentosUpdate, which are monad transformer stacks over the corresponding acid-state types. On top of acid-state access, thentos transactions provide error handling and authorization management.

actions can be composed of more than one acidic transaction in a non-acidic fashion, and possibly do other things involving randomness or the system time. Actions live in the Action monad transformer stack and provide access to acid-state as well as IO, and authentication management. queryAction and updateAction can be used to translate transactions into actions.

A collection of basic transactions is implemented in Thentos.DB.Trans, and one of simple actions is implemented in Thentos.Api. Software using Thentos as a library is expected to add more transactions and actions and place them in other modules.

Authorization

Access to acid-state data is protected by DCLabels from the lio package. Each transaction is called in the context of a clearance level (ThentosClearance) that describes the privileges of the calling user. Each transaction returns its result together with a label (ThentosLabel). The label must satisfy label canFlowTo clearance (see canFlowTo). The functions accessAction, updateAction, queryAction take the clearance level from the application state and attach it to the acid-state event.

Labels can be combined to a least upper bound (method lub of type class Label). If several transactions are combined to a new, more complex transaction, the latter must be labeled with the least upper bound of the more primitive ones.

In order to avoid having to check label against clearance explicitly inside every transaction, we use a Thentos-specific derivative of makeAcidic that calls runThentosUpdate or runThentosQuery, resp.. (FIXME: not implemented yet!)

If you need to implement an action that runs with higher clearance than the current user can present, accessAction takes a Maybe argument to override the presented clearance.

See the (admittedly rather sparse) lio documentation for more details (or ask the thentos maintainers to elaborate on this :-).

type ActionStateGlobal r = (AcidState DB, r, ThentosConfig)

type ActionState r = (ActionStateGlobal r, DB -> TimeStamp -> Either ThentosError ThentosClearance)

type Action r = ReaderT (ActionState r) (EitherT ThentosError IO)

runAction' :: (MonadIO m, CPRG r) => (ActionStateGlobal (MVar r), ThentosClearance) -> Action (MVar r) a -> m (Either ThentosError a)

runAction :: (MonadIO m, CPRG r) => ActionState (MVar r) -> Action (MVar r) a -> m (Either ThentosError a)

catchAction :: Action r a -> (ThentosError -> Action r a) -> Action r a

logActionError :: Action r a -> Action r a

updateAction :: forall event a r. (UpdateEvent event, EventState event ~ DB, EventResult event ~ Either ThentosError a) => (ThentosClearance -> event) -> Action r a

queryAction :: forall event a r. (QueryEvent event, EventState event ~ DB, EventResult event ~ Either ThentosError a) => (ThentosClearance -> event) -> Action r a

addUnconfirmedUser :: CPRG r => UserFormData -> Action (MVar r) (UserId, ConfirmationToken)

addService :: CPRG r => Action (MVar r) (ServiceId, ServiceKey)

startSession :: CPRG r => Agent -> TimeStamp -> Timeout -> Action (MVar r) SessionToken

startSessionNow :: CPRG r => Agent -> Action (MVar r) SessionToken

Sessions have a fixed duration of 2 weeks.

startSessionNowWithTimeout :: CPRG r => (Agent, Timeout) -> Action (MVar r) SessionToken

Sessions with explicit timeout.

bumpSession :: SessionToken -> Action r (SessionToken, Session)

isActiveSession :: SessionToken -> Action r Bool

isActiveSessionAndBump :: SessionToken -> Action r Bool

isLoggedIntoService :: SessionToken -> ServiceId -> Action r Bool

addServiceLogin :: SessionToken -> ServiceId -> Action r ()

dropServiceLogin :: SessionToken -> ServiceId -> Action r ()