Thentos.Backend.Api.Adhocracy3

Description

This is an implementation of git@github.com:liqdadhocracy3.git:docssourceapi/authentication_api.rst

data types

basics

newtype Path

Constructors

Path ST

data ContentType

Constructors

CTUser

data PropertySheet

Constructors

PSUserBasic

PSPasswordAuthentication

resource

data A3Resource a

Constructors

A3Resource (Maybe Path) (Maybe ContentType) (Maybe a)

individual resources

newtype A3UserNoPass

Constructors

A3UserNoPass

Fields

fromA3UserNoPass :: UserFormData

newtype A3UserWithPass

Constructors

A3UserWithPass

Fields

fromA3UserWithPass :: UserFormData

a3UserToJSON :: Bool -> UserFormData -> Value

a3UserFromJSON :: Bool -> Value -> Parser UserFormData

userNameValid :: ST -> Bool

constraints on user name: The "name" field in the IUserBasic schema is a non-empty string that can contain any characters except '@' (to make user names distinguishable from email addresses). The username must not contain any whitespace except single spaces, preceded and followed by non-whitespace (no whitespace at begin or end, multiple subsequent spaces are forbidden, tabs and newlines are forbidden).

FIXME: not implemented.

emailValid :: ST -> Bool

RFC 5322 (sections 3.2.3 and 3.4.1) and RFC 5321

FIXME: not implemented.

passwordGood :: ST -> Bool

Only an empty password is a bad password.

other types

data ActivationRequest

Constructors

ActivationRequest Path

data LoginRequest

Constructors

LoginByName UserName UserPass

LoginByEmail UserEmail UserPass

data RequestResult

Constructors

RequestSuccess Path SessionToken

RequestError [ST]

main

runBackend :: Int -> ActionStateGlobal (MVar SystemRNG) -> IO ()

serveApi :: ActionStateGlobal (MVar SystemRNG) -> Application

api

type App = ("principals" :> ("users" :> (ReqBody A3UserWithPass :> Post (A3Resource A3UserNoPass)))) :<|> (("activate_account" :> (ReqBody ActivationRequest :> Post RequestResult)) :<|> (("login_username" :> (ReqBody LoginRequest :> Post RequestResult)) :<|> (("login_email" :> (ReqBody LoginRequest :> Post RequestResult)) :<|> ServiceProxy)))

Note: login_username and login_email have identical behavior. In particular, it is not an error to send username and password to /login_email. This makes implementing all sides of the protocol a lot easier without sacrificing security.

app :: ActionStateGlobal (MVar SystemRNG) -> Server App

handler

addUser :: A3UserWithPass -> RestAction (A3Resource A3UserNoPass)

activate :: ActivationRequest -> RestAction RequestResult

login :: LoginRequest -> RestAction RequestResult

FIXME: check password!

aux

userIdToPath :: UserId -> Path

userIdFromPath :: Path -> RestAction UserId

confirmationTokenFromPath :: Path -> RestAction ConfirmationToken