Thentos.Backend.Api.Simple

Documentation

type App = ThentosAuth ThentosBasic

data ThentosAuth layout

Empty data type for triggering authentication. If you have an api type API, use like this: ThentosAuth :> API, then write a route handler that takes Auth as an extra argument. Auth will be parsed from the headers and injected into the sublayout handler.

runBackend :: Int -> ActionStateGlobal (MVar SystemRNG) -> IO ()

serveApi :: ActionStateGlobal (MVar SystemRNG) -> Application

(Required in test suite.)