Thentos.Backend.Core

Documentation

type RestAction = Action (MVar SystemRNG)

type RestActionState = ActionState (MVar SystemRNG)

type RestActionRaw = EitherT RestError IO

type RestError = (Int, String)

lookupRequestHeader :: Request -> CI SBS -> Maybe ST

class PushActionC a where

This is a work-around: The Server type family terminates in RestActionRaw on all methods. PushActionC instances transform handlers implemented in a monad stack we want (providing acid state, clearance info, random generator, ... in a reader) into the handlers in RestActionRaw. (Also, translate DbError to RestError.)

Associated Types

type PushActionSubRoute a

Methods

pushAction :: RestActionState -> PushActionSubRoute a -> a

fmapLTM :: (Monad m, Functor m) => (a -> m a') -> EitherT a m b -> EitherT a' m b

Like fmapLT from Data.EitherR, but with the update of the left value constructed in an impure action.