Thentos.Config

Description

FIXME: we need to find a less verbose way to combine command line arguments with data from config file. See also: 1694.

Documentation

getCommand :: FilePath -> IO (Either ConfigError Command)

configLogger :: IO ()

data Command

Constructors

Run ThentosConfig

RunA3 ThentosConfig

ShowDB

data ThentosConfig

Constructors

ThentosConfig

Fields

frontendConfig :: Maybe FrontendConfig

backendConfig :: Maybe BackendConfig

proxyConfig :: Maybe ProxyConfig

smtpConfig :: SmtpConfig

defaultUser :: Maybe (UserFormData, [Role])

data FrontendConfig

Constructors

FrontendConfig

Fields

frontendPort :: Int

data BackendConfig

Constructors

BackendConfig

Fields

backendPort :: Int

data ProxyConfig

Constructors

ProxyConfig

Fields

proxyTargets :: Map ServiceId String

data SmtpConfig

Constructors

SmtpConfig

Fields

smtpSender :: Address

smtpSendmailPath :: FilePath

smtpSendmailArgs :: [String]

emptySmtpConfig :: SmtpConfig

emptyThentosConfig :: ThentosConfig