Thentos.DB.Core

Documentation

newtype ThentosClearance

Constructors

ThentosClearance

Fields

fromThentosClearance :: DCLabel

type ThentosUpdate a = ThentosUpdate' (ThentosLabeled ThentosError) (ThentosLabeled a)

type ThentosUpdate' e a = EitherT e (StateT DB Identity) a

runThentosUpdate :: forall a. Show a => ThentosClearance -> ThentosUpdate a -> Update DB (Either ThentosError a)

FIXME: generalize, so we can use this for both Update and Query. (remove runThentosQuery and ThentosQuery when done.)

type ThentosQuery a = ThentosQuery' (ThentosLabeled ThentosError) (ThentosLabeled a)

type ThentosQuery' e a = EitherT e (ReaderT DB Identity) a

runThentosQuery :: forall a. Show a => ThentosClearance -> ThentosQuery a -> Query DB (Either ThentosError a)

liftThentosQuery :: forall e a. ThentosQuery' e a -> ThentosUpdate' e a

returnDb :: Monad m => ThentosLabel -> a -> EitherT (ThentosLabeled e) m (ThentosLabeled a)

throwDb :: Monad m => ThentosLabel -> ThentosError -> EitherT (ThentosLabeled ThentosError) m (ThentosLabeled a)

thentosPublic :: ThentosLabel

thentosDenied :: ThentosLabel

thentosLabeledPublic :: t -> ThentosLabeled t

thentosLabeledDenied :: t -> ThentosLabeled t

(=%%) :: (ToCNF a, ToCNF b) => a -> b -> ThentosLabel infix 6

data TLMode

Constructors

TLRead

TLReadWrite

makeThentosLabel :: ToCNF a => TLMode -> [a] -> ThentosLabel

See also:

makeClearance_ in Thentos.DB.Protected (not exported)

test cases in TestMain.

restrictThentosLabel :: ToCNF a => TLMode -> a -> ThentosLabel -> ThentosLabel

restrictThentosLabel' :: ToCNF a => TLMode -> Maybe a -> ThentosLabel -> ThentosLabel

makeThentosLabel1 :: ToCNF a => TLMode -> a -> ThentosLabel

makeThentosLabel2 :: (ToCNF a, ToCNF b) => TLMode -> a -> b -> ThentosLabel

makeThentosLabel3 :: (ToCNF a, ToCNF b, ToCNF c) => TLMode -> a -> b -> c -> ThentosLabel

makeThentosLabel4 :: (ToCNF a, ToCNF b, ToCNF c, ToCNF d) => TLMode -> a -> b -> c -> d -> ThentosLabel

makeThentosLabel5 :: (ToCNF a, ToCNF b, ToCNF c, ToCNF d, ToCNF e) => TLMode -> a -> b -> c -> d -> e -> ThentosLabel

createCheckpointLoop :: AcidState st -> Int -> Maybe Int -> IO ThreadId

Create a new thread that calls createCheckpoint synchronously, then waits for timeThreshold miliseconds, then repeats. If sizeThreshold is Just a size, create checkpoint only if size of segment of current change log since last checkpoint is larger than that.

FIXME: check change log size. (i think this is only possible inside acid-state.) https:/github.comacid-state/acid-state.