Thentos.DB.Protect

Documentation

makeThentosClearance :: Maybe ST -> DB -> TimeStamp -> Either ThentosError ThentosClearance

If password cannot be verified, or if only password or only principal is provided, throw an error explaining the problem. If none are provided, set clearance level to allowNothing. If both are provided, look up roles of principal, and set clearance level to that of the principal aka agent and all its roles.

Note: Both Roles and Agents can be used in authorization policies. (User can be used, but it must be wrapped into an UserA.)

allowEverything :: ThentosClearance

Clearance for everything.

allowReadEverything :: ThentosClearance

Clearance to read everything, but write / create nothing.

allowNothing :: ThentosClearance

Clearance only for things labeled thentosPublic.

(*%%) :: (ToCNF a, ToCNF b) => a -> b -> ThentosClearance infix 6

createDefaultUser :: AcidState DB -> Maybe (UserFormData, [Role]) -> IO ()

If default user is Nothing or user with 'UserId 0' exists, do nothing. Otherwise, create default user.