Thentos.DB.Trans

Documentation

newtype AllUserIds

Constructors

AllUserIds ThentosClearance

trans_allUserIds :: ThentosQuery [UserId]

data LookupUser

Constructors

LookupUser UserId ThentosClearance

trans_lookupUser :: UserId -> ThentosQuery (UserId, User)

data LookupUserByName

Constructors

LookupUserByName UserName ThentosClearance

trans_lookupUserByName :: UserName -> ThentosQuery (UserId, User)

data LookupUserByEmail

Constructors

LookupUserByEmail UserEmail ThentosClearance

trans_lookupUserByEmail :: UserEmail -> ThentosQuery (UserId, User)

data AddUser

Constructors

AddUser User ThentosClearance

trans_addUser :: User -> ThentosUpdate UserId

Write new user to DB. Return the fresh user id.

data AddUnconfirmedUser

Constructors

AddUnconfirmedUser ConfirmationToken User ThentosClearance

trans_addUnconfirmedUser :: ConfirmationToken -> User -> ThentosUpdate (UserId, ConfirmationToken)

Write a new unconfirmed user (i.e. one whose email address we haven't confirmed yet) to DB. Unlike addUser, this operation does not ensure uniqueness of email adresses.

data FinishUserRegistration

Constructors

FinishUserRegistration ConfirmationToken ThentosClearance

trans_finishUserRegistration :: ConfirmationToken -> ThentosUpdate UserId

Note on the label for this transaction: If somebody has the confirmation token, we assume that she is authenticated. Since makeThentosClearance does not check for confirmation tokens, this transaction is publicly accessible.

data AddUsers

Constructors

AddUsers [User] ThentosClearance

trans_addUsers :: [User] -> ThentosUpdate [UserId]

Write a list of new users to DB. Return list of fresh user ids. This is not the most vital part of the backend API, but it allows for testing rollback in error cases. It will also be a nice example for intersecting authorizations.

data UpdateUser

Constructors

UpdateUser UserId User ThentosClearance

trans_updateUser :: UserId -> User -> ThentosUpdate ()

Update existing user in DB. Throw an error if user id does not exist, or if email address in updated user is already in use by another user.

data UpdateUserField

Constructors

UpdateUserField UserId UpdateUserFieldOp ThentosClearance

trans_updateUserField :: UserId -> UpdateUserFieldOp -> ThentosUpdate ()

data UpdateUserFieldOp

Constructors

UpdateUserFieldName UserName

UpdateUserFieldEmail UserEmail

UpdateUserFieldAddService ServiceId

UpdateUserFieldDropService ServiceId

data DeleteUser

Constructors

DeleteUser UserId ThentosClearance

trans_deleteUser :: UserId -> ThentosUpdate ()

Delete user with given user id. If user does not exist, throw an error.

newtype AllServiceIds

Constructors

AllServiceIds ThentosClearance

trans_allServiceIds :: ThentosQuery [ServiceId]

data LookupService

Constructors

LookupService ServiceId ThentosClearance

trans_lookupService :: ServiceId -> ThentosQuery (ServiceId, Service)

data AddService

Constructors

AddService ServiceId (HashedSecret ServiceKey) ThentosClearance

trans_addService :: ServiceId -> HashedSecret ServiceKey -> ThentosUpdate ()

Write new service to DB. Service key is generated automatically.

data DeleteService

Constructors

DeleteService ServiceId ThentosClearance

trans_deleteService :: ServiceId -> ThentosUpdate ()

newtype AllSessionTokens

Constructors

AllSessionTokens ThentosClearance

trans_allSessionTokens :: ThentosQuery [SessionToken]

Complete list of all active and inactive sessions. This transaction may go away in the future unless we can think of a good use for it.

data LookupSessionQ

Constructors

LookupSessionQ (Maybe TimeStamp) SessionToken ThentosClearance

trans_lookupSessionQ :: Maybe TimeStamp -> SessionToken -> ThentosQuery (SessionToken, Session)

See trans_lookupSession. The difference is that you cannot bump the session, so there cannot be any changes to the database, so the result monad can be ThentosQuery.

data LookupSession

Constructors

LookupSession (Maybe (TimeStamp, Bool)) SessionToken ThentosClearance

trans_lookupSession :: Maybe (TimeStamp, Bool) -> SessionToken -> ThentosUpdate (SessionToken, Session)

Lookup session. Use first arg to check if session is now active, and if not, remove it from database. Use flag in second arg to update end-of-life TimeStamp in result session and database. If session does not exist, or if active check is enabled and session is inactive, throw NoSuchSession.

Inactive sessions are not deleted. See trans_GarbageCollectSessions.

See trans_lookupSessionQ for a variant of this function in ThentosQuery.

data StartSession

Constructors

StartSession SessionToken Agent TimeStamp Timeout ThentosClearance

trans_startSession :: SessionToken -> Agent -> TimeStamp -> Timeout -> ThentosUpdate SessionToken

Start a new session for user with UserId on service with ServiceId. Start and end time have to be passed explicitly. Throw error if user or service do not exist. Otherwise, return session token. If there is already an active session for this user, return the existing token again, and store new session under it. If agent is a user, remove all her logins (even if session already exists).

data EndSession

Constructors

EndSession SessionToken ThentosClearance

trans_endSession :: SessionToken -> ThentosUpdate ()

End session. Call can be caused by logout request from user (before end of session life time), by session timeouts, or after its natural life time (by application's own garbage collection). If lookup of session owning user fails, throw an error.

data IsActiveSession

Constructors

IsActiveSession TimeStamp SessionToken ThentosClearance

trans_isActiveSession :: TimeStamp -> SessionToken -> ThentosQuery Bool

Is session token currently active? (Do not bump either way.) Never throw errors; if session is inaccessible, simple return False.

Note on authorization: Since session tokens are as powerful as login credentials, this transaction is labelled thentosPublic. (If we wanted to return False for both PermissionDenied and NoSuchSession, we would have to do that in Action rather than in ThentosQuery: PermissionDenied errors can only be constructed from label and clearance, but inside ThentosQuery we have no clearance value.

data IsActiveSessionAndBump

Constructors

IsActiveSessionAndBump TimeStamp SessionToken ThentosClearance

trans_isActiveSessionAndBump :: TimeStamp -> SessionToken -> ThentosUpdate Bool

Is session token currently active? Bump if it is. (See trans_isActiveSession.)

data IsLoggedIntoService

Constructors

IsLoggedIntoService TimeStamp SessionToken ServiceId ThentosClearance

trans_isLoggedIntoService :: TimeStamp -> SessionToken -> ServiceId -> ThentosUpdate Bool

Bump session if it is valid (even if not logged into service).

newtype GarbageCollectSessions

Constructors

GarbageCollectSessions ThentosClearance

trans_garbageCollectSessions :: ThentosQuery [SessionToken]

Go through dbSessions map and find all expired sessions. Return in ThentosQuery. (To reduce database locking, call this and then EndSession on all service ids individually.)

data AssignRole

Constructors

AssignRole Agent Role ThentosClearance

trans_assignRole :: Agent -> Role -> ThentosUpdate ()

Extend Agents entry in dbRoles with a new Role. If Role is already assigned to Agent, do nothing. If Agent does not point to an existing entry in user or service table (or such), throw an error.

data UnassignRole

Constructors

UnassignRole Agent Role ThentosClearance

trans_unassignRole :: Agent -> Role -> ThentosUpdate ()

Extend Agents entry in dbRoles with a new Role. If Role is not assigned to Agent, do nothing. If Agent does not point to an existing entry in user or service table (or such), throw an error.

data LookupAgentRoles

Constructors

LookupAgentRoles Agent ThentosClearance

trans_lookupAgentRoles :: Agent -> ThentosQuery [Role]

All Roles of an Agent. If Agent does not exist or has no entry in dbRoles, return an empty list.

newtype SnapShot

Constructors

SnapShot ThentosClearance

trans_snapShot :: ThentosQuery DB

pure_lookupUser :: DB -> UserId -> Maybe (UserId, User)

pure_lookupUserByName :: DB -> UserName -> Maybe (UserId, User)

pure_lookupService :: DB -> ServiceId -> Maybe (ServiceId, Service)

pure_lookupSession :: DB -> Maybe (TimeStamp, Bool) -> SessionToken -> LookupSessionResult

Lookup session. Use second arg to only return session if now active. Use flag in second arg to update end-of-life TimeStamp in result session.

data LookupSessionResult

Constructors

LookupSessionUnchanged (SessionToken, Session)

LookupSessionBumped (SessionToken, Session)

LookupSessionNotThere

LookupSessionInactive

pure_lookupAgentRoles :: DB -> Agent -> [Role]

emptyDB :: DB

createCheckpointLoop :: AcidState st -> Int -> Maybe Int -> IO ThreadId

Create a new thread that calls createCheckpoint synchronously, then waits for timeThreshold miliseconds, then repeats. If sizeThreshold is Just a size, create checkpoint only if size of segment of current change log since last checkpoint is larger than that.

FIXME: check change log size. (i think this is only possible inside acid-state.) https:/github.comacid-state/acid-state.