Thentos.Frontend.Util

Documentation

serveSnaplet

Arguments

:: Config Snap AppConfig
The configuration of the server - you can usually pass a default Config via defaultConfig.

-> SnapletInit b b
The snaplet initializer function.

-> IO ()

This does the same as serveSnaplet from the snap package, except that it does not try to read app arguments from the command line