Thentos.Types

db

data DB

Constructors

DB

Fields

_dbUsers :: Map UserId User

_dbUnconfirmedUsers :: Map ConfirmationToken (UserId, User)

_dbServices :: Map ServiceId Service

_dbSessions :: Map SessionToken Session

_dbRoles :: Map Agent [Role]

_dbFreshUserId :: !UserId

user

data User

(user groups (the data that services want to store and retrieve in thentos) and session tokens of all active sessions are stored in assoc lists rather than maps. this saves us explicit json instances for now.)

Constructors

User

Fields

_userName :: !UserName

_userPassword :: !(HashedSecret UserPass)

_userEmail :: !UserEmail

_userGroups :: [(ServiceId, [Group])]

_userSession :: !(Maybe SessionToken)

_userLogins :: [ServiceId]

newtype UserId

Constructors

UserId

Fields

fromUserId :: Integer

newtype UserName

Constructors

UserName

Fields

fromUserName :: ST

newtype UserPass

FIXME: ToJSON instance should go away in order to avoid accidental leakage of cleartext passwords. but for the experimentation phase this is too much of a headache. either way, don't do any half-assed rendering to "[password hidden]". causes too many confusing errors.

Constructors

UserPass

Fields

fromUserPass :: ST

newtype HashedSecret a

Constructors

HashedSecret

Fields

fromHashedSecret :: EncryptedPass

newtype UserEmail

Constructors

UserEmail

Fields

fromUserEmail :: ST

newtype Group

Constructors

Group

Fields

fromGroup :: ST

newtype ConfirmationToken

Constructors

ConfirmationToken

Fields

fromConfimationToken :: ST

data UserFormData

Information required to create a new User

Constructors

UserFormData

Fields

udName :: !UserName

udPassword :: !UserPass

udEmail :: !UserEmail

service

data Service

Constructors

Service

Fields

_serviceKey :: !(HashedSecret ServiceKey)

_serviceSession :: !(Maybe SessionToken)

newtype ServiceId

Constructors

ServiceId

Fields

fromServiceId :: ST

newtype ServiceKey

Constructors

ServiceKey

Fields

fromServiceKey :: ST

session, timestamp, timeout

data Session

Constructors

Session

Fields

_sessionAgent :: !Agent

_sessionStart :: !TimeStamp

_sessionEnd :: !TimeStamp

_sessionTimeout :: !Timeout

newtype SessionToken

Constructors

SessionToken

Fields

fromSessionToken :: ST

newtype TimeStamp

Constructors

TimeStamp

Fields

fromTimeStamp :: UTCTime

newtype Timeout

Constructors

Timeout

Fields

fromTimeout :: NominalDiffTime

timeStampToString :: TimeStamp -> String

timeStampFromString :: Monad m => String -> m TimeStamp

timeoutToString :: Timeout -> String

timeoutFromString :: Monad m => String -> m Timeout

role, agent, lio

data Agent

Some thing or body that deals with (and can authenticate itself before) thentos. Examples: User or Service. (We could have called this Principal, but that name is in use by LIO already.)

Constructors

UserA !UserId

ServiceA !ServiceId

data Role

Constructors

RoleAdmin

data ThentosLabeled t

Wrapper for lio's Labeled to avoid orphan instances. (Also, freeze ThentosLabel as label type.)

Constructors

ThentosLabeled

Fields

thentosLabelL :: ThentosLabel

thentosLabelV :: t

newtype ThentosLabel

Wrapper for lio's DCLabel to avoid orphan instances.

Constructors

ThentosLabel

Fields

fromThentosLabel :: DCLabel

newtype ThentosClearance

Constructors

ThentosClearance

Fields

fromThentosClearance :: DCLabel

errors

data ThentosError

Constructors

NoSuchUser

NoSuchPendingUserConfirmation

MalformedConfirmationToken ST

NoSuchService

NoSuchSession

OperationNotPossibleInServiceSession

ServiceAlreadyExists

UserEmailAlreadyExists

UserNameAlreadyExists

PermissionDenied String ThentosClearance ThentosLabel

BadCredentials

BadAuthenticationHeaders

ProxyNotAvailable

MissingServiceHeader

ProxyNotConfiguredForService ServiceId

showThentosError :: MonadIO m => ThentosError -> m (Int, String)

the type of this will change when servant has a better error type.

boilerplate

dbUsers :: Lens' DB (Map UserId User)

dbUnconfirmedUsers :: Lens' DB (Map ConfirmationToken (UserId, User))

dbSessions :: Lens' DB (Map SessionToken Session)

dbServices :: Lens' DB (Map ServiceId Service)

dbRoles :: Lens' DB (Map Agent [Role])

dbFreshUserId :: Lens' DB UserId

userSession :: Lens' User (Maybe SessionToken)

userPassword :: Lens' User (HashedSecret UserPass)

userName :: Lens' User UserName

userLogins :: Lens' User [ServiceId]

userGroups :: Lens' User [(ServiceId, [Group])]

userEmail :: Lens' User UserEmail

sessionTimeout :: Lens' Session Timeout

sessionStart :: Lens' Session TimeStamp

sessionEnd :: Lens' Session TimeStamp

sessionAgent :: Lens' Session Agent

serviceSession :: Lens' Service (Maybe SessionToken)

serviceKey :: Lens' Service (HashedSecret ServiceKey)